Merge size_t_error into main

+++ .src/array_iterator.inl
@@ -5,12 +5,12 @@
 
 namespace jluna
 {
-    template<is_boxable V, size_t R>
-    Array<V, R>::ConstIterator::ConstIterator(size_t i, Array<V, R>* array)
+    template<is_boxable V, uint64_t R>
+    Array<V, R>::ConstIterator::ConstIterator(uint64_t i, Array<V, R>* array)
         : _index(i), _owner(array)
     {}
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     auto& Array<V, R>::ConstIterator::operator++()
     {
         if (_index < _owner->get_n_elements())
@@ -19,7 +19,7 @@ namespace jluna
         return *this;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     auto& Array<V, R>::ConstIterator::operator++(int i)
     {
         if (_index < _owner->get_n_elements())
@@ -28,7 +28,7 @@ namespace jluna
         return *this;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     auto& Array<V, R>::ConstIterator::operator--()
     {
         if (_index > 0)
@@ -37,7 +37,7 @@ namespace jluna
         return *this;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     auto& Array<V, R>::ConstIterator::operator--(int i)
     {
         if (_index > 0)
@@ -46,31 +46,31 @@ namespace jluna
         return *this;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     bool Array<V, R>::ConstIterator::operator==(const typename Array<V, R>::ConstIterator& other) const
     {
         return (this->_owner->operator const jl_value_t *()) == (other._owner->operator const jl_value_t *()) and this->_index == other._index;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     bool Array<V, R>::ConstIterator::operator!=(const typename Array<V, R>::ConstIterator& other) const
     {
         return not (*this == other);
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     auto Array<V, R>::ConstIterator::operator*()
     {
         return Iterator(_index, const_cast<Array<V, R>*>(_owner));
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     auto Array<V, R>::ConstIterator::operator*() const
     {
         return *this;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     Array<V, R>::ConstIterator::operator Proxy()
     {
         gc_pause;
@@ -83,20 +83,20 @@ namespace jluna
         return res;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     template<is_unboxable T, std::enable_if_t<not is < Proxy, T>, bool>>
     Array<V, R>::ConstIterator::operator T() const
     {
         static jl_function_t* getindex = jl_get_function(jl_base_module, "getindex");
-        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<size_t>(_index + 1)));
+        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<uint64_t>(_index + 1)));
     }
 
-    template<is_boxable V, size_t R>
-    Array<V, R>::Iterator::Iterator(size_t i, Array<V, R>* array)
+    template<is_boxable V, uint64_t R>
+    Array<V, R>::Iterator::Iterator(uint64_t i, Array<V, R>* array)
         : ConstIterator(i, array)
     {}
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     template<is_boxable T>
     auto& Array<V, R>::Iterator::operator=(T value)
     {
@@ -106,17 +106,17 @@ namespace jluna
         static jl_function_t* setindex = jl_get_function(jl_base_module, "setindex!");
 
         gc_pause;
-        jl_call3(setindex, _owner->operator jl_value_t *(), box(value), box(_index + 1));
+        jl_call3(setindex, _owner->operator jl_value_t *(), box(value), box((uint64_t)(_index + 1)));
         gc_unpause;
 
         return *this;
     }
 
-    template<is_boxable V, size_t R>
+    template<is_boxable V, uint64_t R>
     template<is_unboxable T, std::enable_if_t<not std::is_same_v<T, Proxy>, bool>>
     Array<V, R>::Iterator::operator T() const
     {
         static jl_function_t* getindex = jl_get_function(jl_base_module, "getindex");
-        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<size_t>(_index + 1)));
+        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<uint64_t>(_index + 1)));
     }
 }
